% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_utils.R
\name{auto_streamflow_params}
\alias{auto_streamflow_params}
\title{Auto-populate streamflow query params from catalog}
\usage{
auto_streamflow_params(dataset_id)
}
\arguments{
\item{dataset_id}{Dataset identifier}
}
\value{
List of recommended params for flowfabric_streamflow_query
}
\description{
Auto-populate streamflow query params from catalog
}
