% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{flowfabric_get}
\alias{flowfabric_get}
\title{Perform a GET request to the FlowFabric API}
\usage{
flowfabric_get(endpoint, token = NULL, ...)
}
\arguments{
\item{endpoint}{API endpoint (e.g., "/v1/datasets")}

\item{token}{Optional. Bearer token. If NULL, will use flowfabric_get_token().}

\item{...}{Additional httr2::request() options}
}
\value{
httr2 response object
}
\description{
Perform a GET request to the FlowFabric API
}
\examples{
\dontrun{
# GET request to a FlowFabric endpoint
resp <- flowfabric_get("/datasets", token = flowfabric_get_token())
}
}
