% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{flowfabric_get_dataset}
\alias{flowfabric_get_dataset}
\title{Get a specific dataset}
\usage{
flowfabric_get_dataset(dataset_id, token = NULL, verbose = FALSE)
}
\arguments{
\item{dataset_id}{Dataset identifier}

\item{token}{Optional. If NULL, will use get_bearer_token()}

\item{verbose}{Use TRUE for debugging purposes}
}
\value{
Returns a specific dataset
}
\description{
Get a specific dataset
}
\examples{
\dontrun{
dataset <- flowfabric_get_dataset("nws_owp_nwm_analysis")
}
}
