% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{flowfabric_get_latest_run}
\alias{flowfabric_get_latest_run}
\title{Get the latest run for a specified dataset}
\usage{
flowfabric_get_latest_run(dataset_id, token = NULL, verbose = FALSE)
}
\arguments{
\item{dataset_id}{Dataset identifier}

\item{token}{Optional. If NULL, will use get_bearer_token()}

\item{verbose}{Use TRUE for debugging purposes}
}
\value{
Returns the latest run for a specified dataset
}
\description{
Get the latest run for a specified dataset
}
\examples{
\dontrun{
latest <- flowfabric_get_latest_run("nws_owp_nwm_analysis")
}
}
