% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{flowfabric_healthz}
\alias{flowfabric_healthz}
\title{Health check}
\usage{
flowfabric_healthz(token = NULL, verbose = FALSE)
}
\arguments{
\item{token}{Optional. Bearer token. Uses get_bearer_token() if NULL.}

\item{verbose}{Use TRUE for debugging purposes}
}
\value{
Health status and metadata
}
\description{
Health check
}
\examples{
\dontrun{
  health <- flowfabric_healthz()
}
}
