% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{flowfabric_inundation_ids}
\alias{flowfabric_inundation_ids}
\title{Query inundation ids}
\usage{
flowfabric_inundation_ids(params = list(), token = NULL, verbose = FALSE)
}
\arguments{
\item{params}{List of query parameters (see API docs)}

\item{token}{Optional Bearer token. If NULL, will use `get_bearer_token()`}

\item{verbose}{Logical; print debug messages}
}
\description{
Wrapper for 'POST /v1/inundation-ids' returning parsed JSON.
}
\examples{
\dontrun{
polygons <- flowfabric_inundation_ids(params=auto_streamflow_params("nws_owp_nwm_analysis"))
}
}
