% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{flowfabric_list_datasets}
\alias{flowfabric_list_datasets}
\title{List all available datasets}
\usage{
flowfabric_list_datasets()
}
\value{
A data.frame of datasets
}
\description{
List all available datasets
}
\examples{
\dontrun{
# List all datasets (uses default token)
datasets <- flowfabric_list_datasets()
# List datasets with explicit token
token <- flowfabric_get_token()
datasets <- flowfabric_list_datasets(token)
}
}
