% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{flowfabric_post}
\alias{flowfabric_post}
\title{Perform a POST request to the FlowFabric API}
\usage{
flowfabric_post(endpoint, body, token = NULL, ..., verbose = FALSE)
}
\arguments{
\item{endpoint}{API endpoint (e.g., "/v1/datasets/nwm-forecast/streamflow")}

\item{body}{List. JSON body for the request.}

\item{token}{Optional. Bearer token. If NULL, will use flowfabric_get_token().}

\item{...}{Additional httr2::request() options}

\item{verbose}{Use TRUE for debugging purposes}
}
\value{
httr2 response object
}
\description{
Perform a POST request to the FlowFabric API
}
\examples{
\dontrun{
# POST request to a FlowFabric endpoint
resp <- flowfabric_post("/datasets", body = list(name = "test"), token = flowfabric_get_token())
}
}
