% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{flowfabric_stage_query}
\alias{flowfabric_stage_query}
\title{Query stage data (Arrow IPC)}
\usage{
flowfabric_stage_query(dataset_id, params = NULL, token = NULL, ...)
}
\arguments{
\item{dataset_id}{Dataset identifier}

\item{params}{List of query parameters (see API docs)}

\item{token}{Optional. Bearer token. If NULL, will use flowfabric_get_token().}
}
\value{
An Arrow Table
}
\description{
Query stage data (Arrow IPC)
}
\examples{
\dontrun{
tbl <- flowfabric_stage_query(
  dataset_id = "usgs-nwis-stage",
  params = list(feature_ids = c("101"))
)
df <- as.data.frame(tbl)
}
}
