% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{flowfabric_streamflow_estimate}
\alias{flowfabric_streamflow_estimate}
\title{Estimate streamflow}
\usage{
flowfabric_streamflow_estimate(
  dataset_id,
  feature_ids = NULL,
  start_time = NULL,
  end_time = NULL,
  issue_time = NULL,
  params = NULL,
  token = NULL,
  ...,
  verbose = FALSE
)
}
\arguments{
\item{dataset_id}{Dataset identifier}

\item{feature_ids}{Optional. Character vector of feature IDs}

\item{start_time}{Optional. Start time of desired data}

\item{end_time}{Optional. End time of desired data}

\item{issue_time}{Optional. Issue time for query parameters}

\item{params}{List of query parameters (see API docs)}

\item{token}{Optional. Bearer token. If NULL, will use flowfabric_get_token()}

\item{...}{Optional. Additional parameters (passed as named list)}

\item{verbose}{Optional. Use TRUE for debugging purposes}
}
\value{
Returns a list
}
\description{
Estimate streamflow
}
\examples{
\dontrun{
result <- flowfabric_streamflow_estimate(dataset_id, params = list(
  query_mode = "run",
  feature_ids = c("101"),
  issue_time = "latest",
  scope = "features",
  lead_start = 0,
  lead_end = 0,
  format = "arrow"
  )
)
}
}
