% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{flowfabric_token}
\alias{flowfabric_token}
\title{Manage global FlowFabric API token}
\usage{
flowfabric_token(refresh = FALSE)
}
\description{
Set or get the global token used for authentication. If not set, will attempt to obtain one using flowfabric_get_token().
}
