% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{flowfabric_get_dataset}
\alias{flowfabric_get_dataset}
\title{Get dataset details from catalog}
\usage{
flowfabric_get_dataset(dataset_id, token = NULL, verbose = FALSE)
}
\arguments{
\item{dataset_id}{Dataset identifier}

\item{token}{Optional Bearer token. If NULL, will use `get_bearer_token()`}

\item{verbose}{Logical; print debug messages}
}
\description{
Wrapper for `GET /v1/datasets/\{dataset_id\}` returning parsed JSON dataset object.
}
\examples{
\dontrun{
dataset <- flowfabric_get_dataset("nws_owp_nwm_analysis")
}
}
