% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{flowfabric_get_run}
\alias{flowfabric_get_run}
\title{Get a specific run for a specified dataset}
\usage{
flowfabric_get_run(dataset_id, issue_time, token = NULL, verbose = FALSE)
}
\arguments{
\item{dataset_id}{Dataset identifier}

\item{issue_time}{Issue time for desired run}

\item{token}{Optional. If NULL, will use get_bearer_token()}

\item{verbose}{Use TRUE for debugging purposes}
}
\value{
Returns a specific run for the specified dataset
}
\description{
Get a specific run for a specified dataset
}
\examples{
\dontrun{
recent_run <- flowfabric_get_run("nws_owp_nwm_analysis", issue_time = "2026010514")
}
}
