% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{flowfabric_streamflow_estimate}
\alias{flowfabric_streamflow_estimate}
\title{Estimate streamflow}
\usage{
flowfabric_streamflow_estimate(
  dataset_id,
  params = NULL,
  token = NULL,
  ...,
  verbose = FALSE
)
}
\arguments{
\item{dataset_id}{Dataset identifier}

\item{params}{List of query parameters (see API docs)}

\item{token}{Optional. Bearer token. If NULL, will use flowfabric_get_token()}

\item{...}{Optional. Additional parameters (passed as named list)}

\item{verbose}{Optional. Use TRUE for debugging purposes}
}
\value{
Returns a list
}
\description{
Estimate streamflow
}
\examples{
\dontrun{
result <- flowfabric_streamflow_estimate(dataset_id, params)
}
}
