% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowfabricr-package.R
\docType{package}
\name{flowfabricr}
\alias{flowfabricr-package}
\alias{flowfabricr}
\title{flowfabricr: R Client for FlowFabric API}
\description{
This package provides a high-performance R interface to the FlowFabric REST API, including JWT authentication (leveraging and improving hfutils). It is designed for expert R users and includes documentation and tests.
}
\author{
\strong{Maintainer}: Mike Johnson \email{mjohnson@lynker.com}

}
\keyword{internal}
