% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_bearer_token}
\alias{get_bearer_token}
\title{Retrieve Bearer Token}
\usage{
get_bearer_token(force_refresh = FALSE)
}
\arguments{
\item{force_refresh}{Will always refresh the token if TRUE}
}
\value{
A character string containing the Bearer token.
}
\description{
This function retrieves a Bearer token using `hfutils::lynker_spatial_auth`.
}
\examples{
\dontrun{
token <- get_bearer_token()
}
}
